Refactor import paths for quiz details screen
